Add IT block check for condition code updates
The S instruction boolean field has been replaced with a trinary
setflags field.  This field specifies whether to always. never, or only
when not in an IT block to update the condition codes.

Add register methods to determine whether we are in an IT block, and
check this when updating condition codes.